@@ -1,11 +1,11 @@
 name: flutter_dtmf
 description: Generates DTFM Tones for Flutter Application. This can be used in VOIP applications or other applications that need to generate DTMF Tones.
-version: 1.0.0
+version: 1.0.1
 homepage: "https://www.dormmom.com"
 repository: "https://github.com/eopeter/flutter_dtmf"
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
